refactor: user 패키지 도메인 생성 및 재이동
