--- src/main/kotlin/dnd11th/blooming/api/service/user/IdTokenValidator.kt
@@ -1,5 +1,0 @@
-package dnd11th.blooming.api.service.user
-
-class IdTokenValidator {
-
-}
