+++ src/main/kotlin/dnd11th/blooming/client/config/FeignClientConfig.kt
@@ -4,5 +4,5 @@ import org.springframework.cloud.openfeign.EnableFeignClients
 import org.springframework.context.annotation.Configuration
 
 @Configuration
-@EnableFeignClients
+@EnableFeignClients(basePackages = ["dnd11th.blooming"])
 class FeignClientConfig
