+++ src/main/kotlin/dnd11th/blooming/client/dto/OidcPublicKey.kt
@@ -6,5 +6,4 @@ data class OidcPublicKey(
     val alg: String,
     val n: String,
     val e: String,
-) {
-}
+)
