+++ src/main/kotlin/dnd11th/blooming/domain/entity/user/UserClaims.kt
@@ -1,7 +1,7 @@
-package dnd11th.blooming.domain.entity
+package dnd11th.blooming.domain.entity.user
 
 data class UserClaims(
-    val id: Long,
+    val id: Long?,
     val email: String,
     val nickname: String,
 )
